server/requirements.txt
Cython==0.21.1
Flask==0.10.1
Flask-Cors==1.10.2
Flask-SQLAlchemy==2.0
Jinja2==2.7.3
Mako==0.8.1
MarkupSafe==0.23
Pillow==2.0.0
PyJWT==0.4.1
SQLAlchemy==0.9.8
Werkzeug==0.9.6
anyjson==0.3.3
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.19
certifi==14.05.14
configglue==1.1.2
defer==1.0.6
dirspec==13.10
distribute==0.6.37
greenlet==0.4.5
httplib2==0.8
itsdangerous==0.24
kombu==3.0.24
oauthlib==0.7.2
pexpect==2.4
piston-mini-client==0.7.5
pyOpenSSL==0.13
pycrypto==2.6
pyinotify==0.9.3
pytz==2014.10
pyxdg==0.25
reportlab==2.6
requests==2.5.1
requests-oauthlib==0.4.2
six==1.8.0
ssh-import-id==3.19
tornado==4.0.2
uWSGI==2.0.8
urllib3==1.6
wsgiref==0.1.2
zope.interface==4.0.5
